IndusInd bank project pyspark databricks code
For any source table data, which includes the Data Quality Framework, SCD-1 and SCD-2 implementations in the Parquet file,.
source from the SQL server of the client to ADLS Gen2 blob storage at the raw layer which is incremental present data, and the archive layer, which has historical data moved from the raw layer. Then, from this storage account, this data is passed to Databricks code from the ADF pipeline, which creates a delta table with the name of the table, which is the data is passed to Databricks code in the Unity catlog, which is called curated-layer.